Get setUser test passing
