+++ src/store.js
@@ -1,5 +1,6 @@
 import { isTestHostname } from "./utils.js";
 
 export const store = {
+  user: {},
   mode: isTestHostname() ? "test" : "live",
 };
