+++ test/user.spec.js
@@ -1,12 +1,12 @@
 import jwt from "jsonwebtoken";
 import axios from "axios";
 
-import utils from "./utils.js";
+import utils from "./config/utils.js";
 import { apiUrl } from "../src/constants.js";
-import createUser from "../src/user.js";
+import { setCookie } from "../src/cookies.js";
 
 import Userfront from "../src/index.js";
-const { user } = Userfront;
+import { setUser } from "../src/user.js";
 
 jest.mock("axios");
 const tenantId = "abcdefgh";
@@ -15,19 +15,12 @@ describe("User", () => {
   beforeAll(async () => {
     // Set the factory access and ID tokens as cookies
     Userfront.store.tenantId = tenantId;
-    Userfront.setCookie(
+    setCookie(
       utils.createAccessToken(),
       { secure: "true", sameSite: "Lax" },
       "access"
     );
-    Userfront.setCookie(
-      utils.createIdToken(),
-      { secure: "true", sameSite: "Lax" },
-      "id"
-    );
-
-    // Mock `Userfront.verifyToken` to prevent implementation
-    Userfront.__set__("verifyToken", () => jest.fn(() => Promise.resolve()));
+    setCookie(utils.createIdToken(), { secure: "true", sameSite: "Lax" }, "id");
 
     // Mock `Userfront.refresh` to assert calls later
     Userfront.__set__("refresh", jest.fn());
@@ -38,27 +31,26 @@ describe("User", () => {
     return Promise.resolve();
   });
 
-  describe("user constructor", () => {
-    it("should set user's information based on ID token and return user object", () => {
-      const parsedUser = {
-        ...JSON.parse(JSON.stringify(utils.idTokenUserDefaults)),
-      };
+  describe("setUser", () => {
+    it("should set user's information based on ID token", () => {
+      const defaultUserValues = utils.idTokenUserDefaults;
 
-      const ufUser = createUser({ store: Userfront.store });
+      // Call setUser
+      setUser();
 
-      // Assert ufUser values were set correctly in user constructor
-      for (const prop in utils.defaultIdTokenProperties) {
-        expect(parsedUser[prop]).toEqual(ufUser[prop]);
+      // Assert primary values were set correctly
+      for (const prop in defaultUserValues) {
+        expect(Userfront.user[prop]).toEqual(defaultUserValues[prop]);
       }
 
-      // Assert ufUser.data values were set correctly
-      for (const prop in parsedUser.data) {
-        expect(parsedUser.data[prop]).toEqual(ufUser.data[prop]);
+      // Assert data values were set correctly
+      for (const prop in defaultUserValues.data) {
+        expect(Userfront.user.data[prop]).toEqual(defaultUserValues.data[prop]);
       }
     });
   });
 
-  describe("user object", () => {
+  xdescribe("user object", () => {
     it("should get user's information", () => {
       const parsedUser = {
         ...JSON.parse(JSON.stringify(utils.idTokenUserDefaults)),
@@ -79,7 +71,7 @@ describe("User", () => {
     });
   });
 
-  describe("user.update()", () => {
+  xdescribe("user.update()", () => {
     it("should update user's information via API then call afterUpdate hook", async () => {
       const updates = {
         username: "john-doe-updated",
